feat 130: fixing and adding new tests
